finish manual e2e tests
